+++ packages/spacecat-shared-user-manager-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@adobe/spacecat-shared-user-manager-client-v1.2.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-user-manager-client-v1.1.0...@adobe/spacecat-shared-user-manager-client-v1.2.0) (2026-06-26)
+
+### Features
+
+* **user-manager-client:** stateful Counterfact mock for the sub-workspace lifecycle gateway (LLMO-5616) ([#1734](https://github.com/adobe/spacecat-shared/issues/1734)) ([50983ac](https://github.com/adobe/spacecat-shared/commit/50983ac45304f2ec94bc5a2d0faa5697adb1b431))
+
 ## [@adobe/spacecat-shared-user-manager-client-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-user-manager-client-v1.0.0...@adobe/spacecat-shared-user-manager-client-v1.1.0) (2026-06-23)
 
 ### Features